+++ src/import-service/utils/constants.js
@@ -1,0 +1,1 @@
+export const BUCKET = 'macbook-shop-uploaded';
